move reading code tables to cdm_mapper.codes
