+++ cdm_reader_mapper/cdm_mapper/codes/codes.py
@@ -9,14 +9,68 @@ Common Data Model (CMD) tables within the cdm tool.
 
 from __future__ import annotations
 
+import ast
+import datetime
+
 from cdm_reader_mapper.common.json_dict import (
     collect_json_files,
     combine_dicts,
-    open_code_table,
+    open_json_file,
 )
 
 from .. import properties
 
+def _eval(s):
+    try:
+      return ast.literal_eval(s)
+    except SyntaxError:
+      return s
+
+def expand_integer_range_key(d):
+        """DOCUMENTATION."""
+        # Looping based on print_nested above
+        if not isinstance(d, dict):
+            return d
+        d_ = {}
+        for k, v in d.items():
+            k_ = _eval(k)
+            if not isinstance(k_ ,list):
+                return {k: expand_integer_range_key(v)}
+            try:
+              lower = int(k_[0])
+            except Exception as e:
+                logging.error(f"Lower bound parsing error in range key: {k}")
+                logging.error(e)
+                return
+            upper = k_[1]
+            if upper == "yyyy":
+                upper = datetime.date.today().year
+            else:
+                try:
+                    upper = int(upper)
+                except Exception as e:
+                    logging.error(f"Lower bound parsing error in range key: {k}")
+                    logging.error(e)
+                    return
+            if len(k_) > 2:
+                try:
+                    step = int(k_[2])
+                except Exception as e:
+                    logging.error(f"Range step parsing error in range key: {k}")
+                    logging.error(e)
+                    return
+            else:
+                step = 1
+                
+            for yr in range(lower, upper + 1, step):
+                d_[str(yr)] = v
+        return d_
+
+
+def _open_code_table(ifile):
+    json_dict = open_json_file(ifile)
+    return expand_integer_range_key(json_dict)
+        
 
 def get_code_table(data_model, *sub_models, code_table=None):
     """Load code tables into dictionary.
@@ -39,5 +93,5 @@ def get_code_table(data_model, *sub_models, code_table=None):
         data_model, *sub_models, base=f"{properties._base}.codes", name=code_table
     )
     table_files = common_files + table_files
-    tables = [open_code_table(ifile) for ifile in table_files]
+    tables = [_open_code_table(ifile) for ifile in table_files]
     return combine_dicts(tables)
